requirements.txt
streamlit
plotly
paho-mqtt
